+++ components/category/categoryList/stories/CategoryList.stories.tsx
@@ -1,45 +1,31 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-
-import CategoryWrapper from '../CategoryWrapper';
-import { ChildrenStory, AnchorStory } from './CategoryWrapper.data';
+import CategoryList from '../CategoryList';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Category/WrapperButton',
-  component: CategoryWrapper,
+  title: 'Category/List',
+  component: CategoryList,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     backgroundColor: { control: 'color' },
   },
-} as ComponentMeta<typeof CategoryWrapper>;
+} as ComponentMeta<typeof CategoryList>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof CategoryWrapper> = (args) => (
-  <CategoryWrapper {...args} />
+const Template: ComponentStory<typeof CategoryList> = (args) => (
+  <CategoryList {...args} />
 );
 
 export const Primary = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
-Primary.args = {
-  anchor: AnchorStory.Default,
-  children: ChildrenStory,
-};
+Primary.args = {};
 
 export const Secondary = Template.bind({});
-Secondary.args = {
-  anchor: AnchorStory.Default,
-  children: ChildrenStory,
-};
+Secondary.args = {};
 
 export const Large = Template.bind({});
-Large.args = {
-  anchor: AnchorStory.Default,
-  children: ChildrenStory,
-};
+Large.args = {};
 
 export const Small = Template.bind({});
-Small.args = {
-  anchor: AnchorStory.Default,
-  children: ChildrenStory,
-};
+Small.args = {};
