+++ components/category/categoryWrapper/__test__/CategoryWrapper.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import CategoryWrapper from '@/components/categoryWrapper/CategoryWrapper';
+import CategoryWrapper from '@/components/category/categoryWrapper/CategoryWrapper';
 
 /**
  * expect extend 먹힘
